Fixed footer shadow.
Fixed the footer elevation shadow so that it now faces upwards. To do
this I had to hard code it, as I could not find any way to swap the
direction of the `mat-elevation` shadow--I doubt it exists.

Because this requires hard coding, it may be worth getting rid of this
shadow or the footer in the future.